Completed all questions of lab 4

+++ PomonaTransitSystem/src/pts/q2/DeleteTripOfferingController.java
@@ -5,6 +5,8 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 
 import javafx.event.ActionEvent;
 import javafx.scene.control.Alert;
@@ -25,30 +27,39 @@ public class DeleteTripOfferingController extends TripOfferingController {
 	public void deleteTripOfferingAction(ActionEvent event) throws SQLException, IOException {
 		try {
 			tripNumber = tripNumberInput.getText();
-			date = dateInput.getText();
+			LocalDate myDate = datePicker.getValue();
 			scheduledStartTime = scheduledStartTimeInput.getText();
 
-			connection = DBConnect.getConnect();
-			query = "DELETE FROM `lab4`.`tripoffering`\r\n"
-					+ "WHERE TripNumber = ? AND Date = ? AND ScheduledStartTime = ?;";
-			preparedStatement = connection.prepareStatement(query);
-			preparedStatement.setString(1, tripNumber);
-			preparedStatement.setString(2, date);
-			preparedStatement.setString(3, scheduledStartTime);
-			preparedStatement.executeUpdate();
-			preparedStatement.close();
-			connection.close();
-			
-			Alert alert = new Alert(AlertType.INFORMATION);
-	        alert.setTitle("Remove Trip Offering");
-	        alert.setHeaderText(null);
-	        alert.setContentText("Trip offering removed successfully!");
-	        alert.showAndWait();
+			if (tripNumber.equals("") || myDate == null || scheduledStartTime.equals("")) {
+				Alert alert = new Alert(AlertType.WARNING);
+				alert.setTitle("Warning");
+				alert.setHeaderText(null);
+				alert.setContentText("Missing data field(s). Please fill out all the fields");
+				alert.showAndWait();
+			} else {
+				date = myDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+				connection = DBConnect.getConnect();
+				query = "DELETE FROM `lab4`.`tripoffering`\r\n"
+						+ "WHERE TripNumber = ? AND Date = ? AND ScheduledStartTime = ?;";
+				preparedStatement = connection.prepareStatement(query);
+				preparedStatement.setString(1, tripNumber);
+				preparedStatement.setString(2, date);
+				preparedStatement.setString(3, scheduledStartTime);
+				preparedStatement.executeUpdate();
+				preparedStatement.close();
+				connection.close();
 
-			tripNumberInput.clear();
-			dateInput.clear();
-			scheduledStartTimeInput.clear();
-			loadTripOffering();
+				Alert alert = new Alert(AlertType.INFORMATION);
+				alert.setTitle("Remove Trip Offering");
+				alert.setHeaderText(null);
+				alert.setContentText("Trip offering removed successfully!");
+				alert.showAndWait();
+
+				tripNumberInput.clear();
+				datePicker.setValue(null);
+				scheduledStartTimeInput.clear();
+				loadTripOffering();
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 			Alert alert = new Alert(AlertType.ERROR);
